Rework node discovery, creation and updating (#7)
This commit refactors the code of core.py to isolate functionality.
Though that is not the main concern with this commit.

This commit also takes on the problem described in issue #7.
Nodes that do not have a node definition in the graph description are
now imported without failure.

The approach taken is to create a node with the dummy label
"prov2neo:node" such that a node exists even if there is no further
information attached to it. The dummy label is appended to each node
that is imported using prov2neo. If later on a solution is found, that
does not add information to the imported PROV graph,
this solution will be abandoned.

The second part of this approach, is to change the transactional
import to try to merge nodes that are in the database with the
ones that get imported rather than to create new nodes every
time (and fail sometimes). This enables the batched graph
imports that are described in #7.

This commit therefore implements the ability to import a provenance
graph in multiple subgraph batches and allows to import graphs
into databases that already contain an arbitrary amount of graphs.