Rework edge conversion (fix #8)
This commit reworks edge conversion and fixes issue #8.

Previously the wrong start and end point of edges could be chosen.
This has now been fixed as the start and end point are always the first
two entries of the formal attributes of a relation.

All other entries in the formal and extra attributes
are now converted according to the attribute encoding and are added as
relation attribute/property entries.

This commit also adds the method Importer.get_node which is used to
retrieve a node by its id. If the node does not exist at the time of the
method call, the node is created within the method and is added to
self.node_dict.