Add exceptions and scheme handling to `client.py` #13
Adds a class for connection errors that occur when trying to connect to the neo4j service instance.

Adds a class for connection schemes that checks whether the scheme is supported. This check is not necessary for the command line version, as the supported schemes are already checked when the command line arguments are parsed. This check is for the python lib.

Outsources the database creation to a standalone method. ClientErrors and DatabaseErrors are raised when necessary.